+++ Prototype/UserControl2.cs
@@ -45,34 +45,38 @@ namespace Prototype
 
         private void removeButton_Click(object sender, EventArgs e)
         {
-            string query = "DELETE FROM image_list WHERE Name = @name";
-
-
-            using (SqlConnection conn = new SqlConnection(connectionString))
-            using (SqlCommand deletcmd = new SqlCommand(query, conn))
+            DialogResult dialogResult = MessageBox.Show($"Are you sure you want to Delete  Car {name}?", "Confirm", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (dialogResult == DialogResult.Yes)
             {
-                deletcmd.Parameters.AddWithValue("@name", name);
-                try
-                {
-                    conn.Open();
-                    int rowsAffected = deletcmd.ExecuteNonQuery();
-                    Console.WriteLine($"Deleted {rowsAffected} row(s)!");
-                    MessageBox.Show($"The {name} remove from shop");
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine($"Error: {ex.Message}");
-                    MessageBox.Show($"The {name} didn't remove from shop");
+                string query = "DELETE FROM image_list WHERE Name = @name";
 
+
+                using (SqlConnection conn = new SqlConnection(connectionString))
+                using (SqlCommand deletcmd = new SqlCommand(query, conn))
+                {
+                    deletcmd.Parameters.AddWithValue("@name", name);
+                    try
+                    {
+                        conn.Open();
+                        int rowsAffected = deletcmd.ExecuteNonQuery();
+                        Console.WriteLine($"Deleted {rowsAffected} row(s)!");
+                        MessageBox.Show($"The {name} remove from shop");
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.WriteLine($"Error: {ex.Message}");
+                        MessageBox.Show($"The {name} didn't remove from shop");
+
+                    }
                 }
             }
+
+                
         }
 
         private void updatePriceButton_Click(object sender, EventArgs e)
         {
-            
-
-
+         
             UpdateCar.price = int.Parse(price, NumberStyles.Currency);
             UpdateCar.name = name;
             UpdateCar.image = image;
